requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
anthropic>=0.18.0
claude-api>=1.0.17
claude-code-sdk>=0.1.0
anyio>=4.0.0 